Métodos de validación y excepciones personalizadas en Cargo

+++ src/main/java/com/cordillera/application/service/CargoService.java
@@ -3,13 +3,14 @@ package com.cordillera.application.service;
 import com.cordillera.application.mapper.CargoMapper;
 import com.cordillera.application.repository.jpa.CargoRepository;
 import com.cordillera.domain.dto.CargoDto;
+import com.cordillera.domain.excepcion.CargoException;
 import com.cordillera.domain.models.Cargo;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -17,43 +18,53 @@ public class CargoService {
 
     //TODO AJUSTAR LA INYECCION DE DEPENDCIA SI POR CONTRUCTOR O POR CAMPOS
     //TODO https://http.cat/
-    @Autowired
     private final CargoMapper cargoMapper;
-
-    @Autowired
     private final CargoRepository cargoRepository;
 
     //Metodo para guardar un cargo
-    public CargoDto saveCargo(CargoDto cargoDto){
+    public CargoDto saveCargo(CargoDto cargoDto) throws CargoException{
         //TODO VALIDACIONES
         //TODO VALIDAR CARGO NO EXISTA
         //TODO METODO NO HACE NADA GUARDAR EN BASE DE DATOS
         //TODO realizar excepcion personalizada
-        return cargoDto;
+        Cargo cargoValidacionNombre = cargoRepository.findCargoByNombreCargo(cargoDto.getNombreCargo()).orElse(null);
+        if(cargoValidacionNombre != null){
+            throw new CargoException("Cargo ya existe en el sistema");
+        }
+        return cargoMapper.cargoModelToCargoDTO(cargoRepository.save(cargoMapper.cargoDTOToCargoModel(cargoDto)));
     }
 
     //Metodo para actualizar un cargo
-    public CargoDto updateCargo(CargoDto cargoDto){
+    public CargoDto updateCargo(CargoDto cargoDto) throws CargoException{
         //TODO VALIDACIONES
         //TODO VALIDAR DESCRIOCION NO EXISTA
         //TODO realizar excepcion personalizada
+        Cargo cargoValidacionNombre = cargoRepository.findCargoByNombreCargo(cargoDto.getNombreCargo()).orElse(null);
+        if(cargoValidacionNombre == null){
+            throw new CargoException("No se puede tener cargos sin nombre");
+        }
         return cargoMapper.cargoModelToCargoDTO(cargoRepository.save(cargoMapper.cargoDTOToCargoModel(cargoDto)));
     }
 
     //Metodo para eliminar un cargo
-    public void deleteCargo(Long cargoId){
+    public void deleteCargo(Long cargoId) throws CargoException{
         //TODO VALIDACIONES REGISTRO BORRAR EXISTA
         //TODO realizar excepcion personalizada
+        Cargo cargoValidacionId = cargoRepository.findById(cargoId).orElse(null);
+        if(cargoValidacionId == null){
+            throw new CargoException("Debe seleccionar al menos un cargo");
+        }
         cargoRepository.deleteById(cargoId);
     }
 
     //Metodo para listar un cargo
-    public List<CargoDto> listCargo(){
+    public List<CargoDto> listCargo() throws CargoException{
         List<Cargo> listaCargos = cargoRepository.findAll();
         List<CargoDto> cargosDTO = new ArrayList<>();
 
         if(listaCargos.isEmpty()){
             //TODO realizar excepcion personalizada
+            throw new RuntimeException("No hay cargos en el sistema");
         }
 
         for(Cargo lCargo: listaCargos){
@@ -65,6 +76,20 @@ public class CargoService {
     }
 
     //TODO METODO PARA CONSULTAR POR ID
+    public CargoDto findById(Long cargoId) throws CargoException{
+        Cargo cargoValidacionId = cargoRepository.findById(cargoId).orElse(null);
+        if(cargoValidacionId == null){
+            throw new CargoException("Cargo no existe en el sistema");
+        }
+        return cargoMapper.cargoModelToCargoDTO(cargoRepository.findById(cargoId).get());
+    }
 
     //TODO METODO PARA CONSULTAR POR DESCRIPCION
+    public CargoDto findByNombre(String nombreCargo) throws CargoException{
+        Cargo cargoValidacionNombre = cargoRepository.findCargoByNombreCargo(nombreCargo).orElse(null);
+        if(cargoValidacionNombre == null){
+            throw new CargoException("El campo nombre se encuentra vacio o no existe en el sistema");
+        }
+        return cargoMapper.cargoModelToCargoDTO(cargoRepository.findCargoByNombreCargo(nombreCargo).get());
+    }
 }
